Moved code to examples
