+++ examples/heating_and_cooling/model/Heating and cooling network with return network.esdl
