--- tests/models/wko/input/timeseries_3.csv
@@ -1,10 +1,0 @@
-DateTime,HeatingDemand_1,HeatingDemand_2,CoolingDemand_1
-01-01-2019 0:00,500000.,500000.,100000.
-01-01-2019 1:00,500000.,500000.,100000.
-01-01-2019 2:00,100000.,100000.,200000.
-01-01-2019 3:00,100000.,100000.,200000.
-01-01-2019 4:00,1000000.,1000000.,0.
-01-01-2019 5:00,2500000.,2500000.,0.
-01-01-2019 6:00,2500000.,2500000.,0.
-01-01-2019 7:00,2500000.,2500000.,0.
-01-01-2019 8:00,0.,0.,1500000.
